Added modal menu items - task #809
