+++ src/MenuBuilder/MenuItemLinkButtonModal.php
@@ -2,31 +2,7 @@
 
 namespace Menu\MenuBuilder;
 
-class MenuItemLinkButtonModal extends BaseMenuItem
+class MenuItemLinkButtonModal extends MenuItemLinkModal
 {
-    /**
-     * @var $modalTarget
-     */
-    protected $modalTarget = null;
 
-    /**
-     * getModalTarget method
-     *
-     * @return string modal target ID
-     */
-    public function getModalTarget()
-    {
-        return $this->modalTarget;
-    }
-
-    /**
-     * setModalTarget method
-     *
-     * @param string $modalTarget ID
-     * @return void
-     */
-    public function setModalTarget($modalTarget)
-    {
-        $this->modalTarget = $modalTarget;
-    }
 }
